+++ modules/nodegroup/variables.tf
@@ -1,0 +1,302 @@
+variable "prefix" {
+  description = "The prefix name of customer to be displayed in AWS console and resource"
+  type        = string
+}
+
+variable "environment" {
+  description = "To manage a resources with tags"
+  type        = string
+}
+
+variable "tags" {
+  description = "A map of tags to add to all resources"
+  type        = map(string)
+  default     = {}
+}
+################################################################################
+# node group
+################################################################################
+variable "node_role_arn" {
+  description = "Amazon Resource Name (ARN) of the IAM Role that provides permissions for the EKS Node Group."
+  type        = string
+}
+
+variable "instance_types" {
+  description = "List of instance types associated with the EKS Node Group"
+  type        = list(string)
+  default     = ["t3.medium"]
+}
+
+variable "labels" {
+  description = "Key-value map of Kubernetes labels. Only labels that are applied with the EKS API are managed by this argument. Other Kubernetes labels applied to the EKS Node Group will not be managed"
+  type        = map(any)
+  default     = {}
+}
+
+variable "subnet_ids" {
+  description = "Identifiers of EC2 Subnets to associate with the EKS Node Group"
+  type        = list(string)
+}
+
+variable "ami_type" {
+  description = " Type of Amazon Machine Image (AMI) associated with the EKS Node Group"
+  type        = string
+  default     = null
+}
+
+variable "is_spot_instances" {
+  description = "Type of capacity associated with the EKS Node Group true if spot_instances else for ondemand"
+  type        = bool
+  default     = false
+}
+
+variable "disk_size" {
+  description = "Disk size in GiB for worker nodes"
+  type        = number
+  default     = 20
+}
+
+variable "desired_size" {
+  description = "Desired number of worker nodes"
+  type        = number
+  default     = 1
+}
+
+variable "max_size" {
+  description = "Maximum number of worker nodes."
+  type        = number
+  default     = 1
+}
+
+variable "min_size" {
+  description = "Minimum number of worker nodes."
+  type        = number
+  default     = 1
+}
+
+variable "max_unavailable" {
+  description = "Desired max number of unavailable worker nodes during node group update."
+  type        = number
+  default     = 1
+}
+
+variable "taint" {
+  description = "The Kubernetes taints to be applied to the nodes in the node group. Maximum of 50 taints per node group"
+  type        = list(any)
+  default     = []
+}
+
+variable "name" {
+  description = "The Name of the nodegroup and luanch teamplate"
+  type        = string
+}
+
+variable "platform" {
+  description = "Identifies if the OS platform is `linux` (currently support only linux)"
+  type        = string
+  default     = "linux"
+}
+
+################################################################################
+# User Data
+################################################################################
+
+variable "enable_bootstrap_user_data" {
+  description = "Determines whether the bootstrap configurations are populated within the user data template"
+  type        = bool
+  default     = true
+}
+
+variable "cluster_name" {
+  description = "Name of associated EKS cluster"
+  type        = string
+  default     = null
+}
+
+variable "cluster_endpoint" {
+  description = "Endpoint of associated EKS cluster"
+  type        = string
+  default     = ""
+}
+
+variable "cluster_auth_base64" {
+  description = "Base64 encoded CA of associated EKS cluster"
+  type        = string
+  default     = ""
+}
+
+variable "cluster_service_ipv4_cidr" {
+  description = "The CIDR block to assign Kubernetes service IP addresses from. If you don't specify a block, Kubernetes assigns addresses from either the 10.100.0.0/16 or 172.20.0.0/16 CIDR blocks"
+  type        = string
+  default     = null
+}
+
+variable "pre_bootstrap_user_data" {
+  description = "User data that is injected into the user data script ahead of the EKS bootstrap script. Not used when `platform` = `bottlerocket`"
+  type        = string
+  default     = ""
+}
+
+variable "post_bootstrap_user_data" {
+  description = "User data that is appended to the user data script after of the EKS bootstrap script. Not used when `platform` = `bottlerocket`"
+  type        = string
+  default     = ""
+}
+
+variable "bootstrap_extra_args" {
+  description = "Additional arguments passed to the bootstrap script. When `platform` = `bottlerocket`; these are additional [settings](https://github.com/bottlerocket-os/bottlerocket#settings) that are provided to the Bottlerocket user data"
+  type        = string
+  default     = ""
+}
+
+################################################################################
+# Launch template
+################################################################################
+
+variable "is_create_launch_template" {
+  description = "Determines whether to create a launch template or not. If set to `false`, EKS will use its own default launch template"
+  type        = bool
+  default     = false
+}
+
+variable "ebs_optimized" {
+  description = "If true, the launched EC2 instance(s) will be EBS-optimized"
+  type        = bool
+  default     = null
+}
+
+variable "ami_id" {
+  description = "The AMI from which to launch the instance. If not supplied, EKS will use its own default image"
+  type        = string
+  default     = ""
+}
+
+variable "key_name" {
+  description = "The key name that should be used for the instance(s)"
+  type        = string
+  default     = null
+}
+
+variable "vpc_security_group_ids" {
+  description = "A list of security group IDs to associate"
+  type        = list(string)
+  default     = []
+}
+
+variable "launch_template_default_version" {
+  description = "Default version of the launch template"
+  type        = string
+  default     = null
+}
+
+variable "update_launch_template_default_version" {
+  description = "Whether to update the launch templates default version on each update. Conflicts with `launch_template_default_version`"
+  type        = bool
+  default     = true
+}
+
+variable "disable_api_termination" {
+  description = "If true, enables EC2 instance termination protection"
+  type        = bool
+  default     = null
+}
+
+variable "kernel_id" {
+  description = "The kernel ID"
+  type        = string
+  default     = null
+}
+
+variable "ram_disk_id" {
+  description = "The ID of the ram disk"
+  type        = string
+  default     = null
+}
+
+variable "block_device_mappings" {
+  description = "Specify volumes to attach to the instance besides the volumes specified by the AMI"
+  type        = any
+  default     = {}
+}
+
+variable "capacity_reservation_specification" {
+  description = "Targeting for EC2 capacity reservations"
+  type        = any
+  default     = null
+}
+
+variable "cpu_options" {
+  description = "The CPU options for the instance"
+  type        = map(string)
+  default     = null
+}
+
+variable "credit_specification" {
+  description = "Customize the credit specification of the instance"
+  type        = map(string)
+  default     = null
+}
+
+variable "elastic_gpu_specifications" {
+  description = "The elastic GPU to attach to the instance"
+  type        = map(string)
+  default     = null
+}
+
+variable "elastic_inference_accelerator" {
+  description = "Configuration block containing an Elastic Inference Accelerator to attach to the instance"
+  type        = map(string)
+  default     = null
+}
+
+variable "enclave_options" {
+  description = "Enable Nitro Enclaves on launched instances"
+  type        = map(string)
+  default     = null
+}
+
+variable "instance_market_options" {
+  description = "The market (purchasing) option for the instance"
+  type        = any
+  default     = null
+}
+
+variable "license_specifications" {
+  description = "A list of license specifications to associate with"
+  type        = map(string)
+  default     = null
+}
+
+variable "metadata_options" {
+  description = "Customize the metadata options for the instance"
+  type        = map(string)
+  default = {
+    http_endpoint               = "enabled"
+    http_tokens                 = "required"
+    http_put_response_hop_limit = 2
+  }
+}
+
+variable "enable_monitoring" {
+  description = "Enables/disables detailed monitoring"
+  type        = bool
+  default     = true
+}
+
+variable "network_interfaces" {
+  description = "Customize network interfaces to be attached at instance boot time"
+  type        = list(any)
+  default     = []
+}
+
+variable "placement" {
+  description = "The placement of the instance"
+  type        = map(string)
+  default     = null
+}
+
+variable "launch_template_tags" {
+  description = "A map of additional tags to add to the tag_specifications of launch template created"
+  type        = map(string)
+  default     = {}
+}